fix: keep importance weights finite and weight the eval estimator
Exponentiate log likelihood ratios in double precision because a
long-horizon tilt can push the exponent past the single-precision
range, fall back to the unweighted quantile when a degenerate batch
sums its weights to zero rather than dividing by it, and give the
evaluation expected-shortfall estimator a weights parameter with a
weighted quantile and tail mean, because evaluating a tilted sample
without its ratios silently estimates the wrong measure.

+++ src/deephedging/training/trainer.py
@@ -82,7 +82,7 @@ def _importance_weights(state: MarketState) -> torch.Tensor | None:
     log_weight = state.aux.get(LOG_WEIGHT_CHANNEL)
     if log_weight is None:
         return None
-    return torch.exp(log_weight)
+    return torch.exp(log_weight.to(torch.float64))
 
 
 class _EpisodeLoss(torch.nn.Module):
